# openapi.yaml
openapi: 3.1.0
info:
  title: Sapliy API
  description: Official API for the Sapliy Ecosystem. Build robust financial applications with our type-safe, high-integrity platform.
  version: 1.0.0
  contact:
    name: Sapliy Support
    url: https://sapliy.io/support
    email: support@sapliy.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sapliy.io
    description: Production API
  - url: http://localhost:8080
    description: Local Development (Gateway)
components:
  parameters:
    ZoneIdHeader:
      name: X-Zone-ID
      in: header
      required: true
      schema:
        type: string
      description: The ID of the zone for this request.
    ZoneModeHeader:
      name: X-Zone-Mode
      in: header
      required: false
      schema:
        type: string
        enum: [live, test]
      description: The mode of the zone (live or test).
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Use your secret or publishable API key as a Bearer token.
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
      description: Session token used for dashboard and frontend requests.
  responses:
    BadRequest:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: RATE_LIMIT_EXCEEDED
            message:
              type: string
              example: Rate limit exceeded. Retry after 60 seconds.
            request_id:
              type: string
            trace_id:
              type: string
    PaymentIntent:
      type: object
      required: [id, amount, currency, status]
      properties:
        id:
          type: string
          example: pi_12345
        amount:
          type: integer
          format: int64
          example: 2000
        currency:
          type: string
          example: USD
        status:
          type: string
          enum:
            [
              requires_payment_method,
              requires_confirmation,
              requires_action,
              processing,
              requires_capture,
              canceled,
              succeeded,
            ]
        client_secret:
          type: string
          description: Used for client-side confirmation.
        description:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
...
    AutomationFlow:
      type: object
      required: [id, name, zone_id]
      properties:
        id:
          type: string
        org_id:
          type: string
        zone_id:
          type: string
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        version:
          type: integer
        trigger:
          $ref: "#/components/schemas/AutomationFlowTrigger"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/AutomationFlowNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/AutomationFlowEdge"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AutomationFlowTrigger:
      type: object
      required: [type]
      properties:
        type:
          type: string
        event_type:
          type: string
        config:
          type: object
    AutomationFlowNode:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
        position:
          type: object
        data:
          type: object
    AutomationFlowEdge:
      type: object
      required: [id, source, target]
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        source_handle:
          type: string
...
paths:
...
  /v1/payments:
    post:
      summary: Create a Payment Intent
      operationId: createPaymentIntent
      tags: [Payments]
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ZoneIdHeader"
        - $ref: "#/components/parameters/ZoneModeHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, currency]
              properties:
                amount:
                  type: integer
                  format: int64
                currency:
                  type: string
                description:
                  type: string
                metadata:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntent"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /v1/events/emit:
    post:
      summary: Emit an Event
      operationId: emitEvent
      tags: [Events]
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                data:
                  type: object
                idempotency_key:
                  type: string
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  event_id:
                    type: string
                  status:
                    type: string
...
  /v1/flows:
    post:
      summary: Create a Flow
      operationId: createFlow
      tags: [Flows]
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationFlow"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationFlow"
...
  /v1/zones:
    post:
      summary: Create a Zone
      operationId: createZone
      tags: [Zones]
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org_id, name, mode]
              properties:
                org_id:
                  type: string
                name:
                  type: string
                mode:
                  type: string
                  enum: [live, test]
                template_name:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  org_id:
                    type: string
                  name:
                    type: string
                  mode:
                    type: string
    get:
      summary: List or Get Zones
      operationId: listZones
      tags: [Zones]
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - name: org_id
          in: query
          schema:
            type: string
        - name: id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    org_id:
                      type: string
                    name:
                      type: string
                    mode:
                      type: string